[py2py3] fix some pylint --py3k warnings - #10440 - src
- src/python/WMCore/WebTools/Masthead.py
- src/python/WMCore/REST/Server.py
- src/python/WMCore/Agent/Database/DestroyAgentBase.py
- src/python/WMCore/ResourceControl/ResourceControl.py

round()->int(round()) - #10440 - src

- src/python/WMCore/Services/LogDB/LogDBBackend.py
- src/python/WMCore/JobSplitting/EventAwareLumiByWork.py

+++ src/python/WMCore/JobSplitting/EventAwareLumiByWork.py
@@ -324,7 +324,7 @@ class EventAwareLumiByWork(JobFactory):
                     for lumi in runLumi.lumis:
                         if (runLumi.run, lumi) in lumiList:
                             self.filesByLumi[run][lumi].append(fileObj)
-                            eventsByLumi[run][lumi] = round(eventsInFile / lumisInFile)
+                            eventsByLumi[run][lumi] = int(round(eventsInFile / lumisInFile))
 
         return lumisByFile, eventsByLumi
